MERGEFIX

+++ WDL/IPlug/IPlugStructs.h
@@ -235,7 +235,7 @@ struct IRECT
 
   inline IRECT SubRectVertical(int numSlices, int sliceIdx)
   {
-    float heightOfSubRect = (float(H()) / numSlices);
+    int heightOfSubRect = int(float(H()) / numSlices);
     int t = heightOfSubRect * sliceIdx;
 
     return IRECT(L, T + t, R, T + t + heightOfSubRect);
@@ -243,7 +243,7 @@ struct IRECT
 
   inline IRECT SubRectHorizontal(int numSlices, int sliceIdx)
   {
-    float widthOfSubRect = (float(W()) / numSlices);
+    int widthOfSubRect = int(float(W()) / numSlices);
     int l = widthOfSubRect * sliceIdx;
 
     return IRECT(L + l, T, L + l + widthOfSubRect, B);
